requirements.txt
flask
psycopg2-binary
python-dotenv
json5
flask-cors